pcs crs from deterministic seed in testing across mpi
