+++ Cargo.toml
@@ -20,7 +20,6 @@ members = [
 ]
 resolver = "2"
 
-
 [workspace.dependencies]
 ark-std = "0.4"
 ark-bn254 = "0.4.0"
@@ -46,3 +45,4 @@ tynm = { version = "0.1.6", default-features = false }
 warp = "0.3.7"
 thiserror = "1.0.63"
 ethnum = "1.5.0"
+rand_chacha = "0.3.1"
