+++ gkr/Cargo.toml
@@ -28,6 +28,7 @@ sha2.workspace = true
 halo2curves.workspace = true
 thiserror.workspace = true
 ethnum.workspace = true
+rand_chacha.workspace = true
 
 # for the server
 bytes.workspace = true
